Fix incorrect field of `faces` in `Polyhedron` schema

+++ src/cdpyr/schema/geometry/polyhedron.py
@@ -9,19 +9,18 @@ __email__ = "p.tempel@tudelft.nl"
 
 class PolyhedronSchema(_geometry.PrimitiveSchema):
     vertices = fields.List(
-            fields.Tuple(
-                    (fields.Float(required=True),
-                     fields.Float(required=True),
-                     fields.Float(required=True))
-            ),
+            fields.Tuple((fields.Float(required=True),
+                          fields.Float(required=True),
+                          fields.Float(required=True))
+                         ),
             required=True,
     )
+
     faces = fields.List(
-            fields.Tuple(
-                    (fields.Float(required=True),
-                     fields.Float(required=True),
-                     fields.Float(required=True))
-            ),
+            fields.Tuple((fields.Integer(required=True),
+                          fields.Integer(required=True),
+                          fields.Integer(required=True))
+                         ),
             required=True,
     )
